requirements.txt
torch>=1.9.0
numpy>=1.21.0
pyyaml>=5.4.0
requests>=2.28.0

sympy>=1.9.0
openai>=1.0.0

# z3-solver>=4.8.0
# networkx>=2.6.0
# albumentations>=1.1.0
# langchain>=0.0.200
# pandas>=1.3.0
# jinja2>=3.0.0
